Performed requested changes for merge into develop branch (2/2)
