+++ pycqed/instrument_drivers/physical_instruments/_CCL/ccl_param_nodes.json
@@ -385,10 +385,10 @@
         }
     ],
     "version": {
-        "Embedded Software Build Time": "02/08/2018-04:07:56",
-        "Embedded Software Version": "0.4.0",
+        "Embedded Software Build Time": "12/10/2018-12:57:11",
+        "Embedded Software Version": "0.5.0",
         "Firmware Build Time": "23/08/2018-15:00:09",
-        "Kernel Module Build Time": "02/08/2018-15:19:15",
+        "Kernel Module Build Time": "06/09/2018-11:17:30",
         "Kernel Module Version": "0.4.0",
         "firmware": "0.7.5",
         "model": "CCL1",
